Обновлена версия Добавлена обработка большей части ошибок Фильтрация ошибок введена в библиотеки

+++ python_rucaptcha/ImageCaptcha.py
@@ -63,7 +63,11 @@ class ImageCaptcha:
             captcha_id = (requests.request('POST',
                                             url_request,
                                             data=payload,
-                                            files=files).json())['request']
+                                            files=files).json())
+        if captcha_id['status'] is 0:
+            return RuCaptchaError(captcha_id['request'])
+
+        captcha_id = captcha_id['request']
 
         # удаляем файл капчи и врменные файлы
         os.remove(os.path.join(self.img_path, "im-{0}.jpg".format(image_hash)))
@@ -75,9 +79,11 @@ class ImageCaptcha:
             captcha_response = requests.request('GET',
                                                 url_response+"?key={0}&action=get&id={1}&json=1"
                                                 .format(self.RUCAPTCHA_KEY, captcha_id))
-            if captcha_response.json()["request"] == 'CAPCHA_NOT_READY':
+            if captcha_response.json()['request']=='CAPCHA_NOT_READY':
                 time.sleep(self.sleep_time)
-            else:
+            elif captcha_response.json()["status"]==0:
+                return RuCaptchaError(captcha_response.json()["request"])
+            elif captcha_response.json()["status"]==1 :
                 return captcha_response.json()['request']
 
     def __del__(self):
@@ -85,4 +91,3 @@ class ImageCaptcha:
             shutil.rmtree(".cache")
         if os.path.exists(self.img_path):
             shutil.rmtree(self.img_path)
-
