+++ python_rucaptcha/MediaCaptcha.py
@@ -77,10 +77,15 @@ class MediaCaptcha:
 
             # Отправляем на рукапча аудио капчи и другие парметры,
             # в результате получаем JSON ответ с номером решаемой капчи и получая ответ - извлекаем номер
-            captcha_id = (requests.request('POST',
+            captcha_id = requests.request('POST',
                                            url_request,
                                            data=payload,
-                                           files=files).json())['request']
+                                           files=files).json()
+        # Фильтрация ошибки
+        if captcha_id['status'] is 0:
+            return RuCaptchaError(captcha_id['request'])
+
+        captcha_id = captcha_id['request']
 
         # удаляем файл капчи и врменные файлы
         os.remove(os.path.join(self.audio_path, 'aud-{0}.mp3'.format(audio_hash)))
@@ -92,9 +97,11 @@ class MediaCaptcha:
             captcha_response = requests.request('GET',
                                                 url_response + "?key={0}&action=get&id={1}&json=1"
                                                 .format(self.RUCAPTCHA_KEY, captcha_id))
-            if captcha_response.json()["request"] == 'CAPCHA_NOT_READY':
+            if captcha_response.json()['request']=='CAPCHA_NOT_READY':
                 time.sleep(self.sleep_time)
-            else:
+            elif captcha_response.json()["status"]==0:
+                return RuCaptchaError(captcha_response.json()["request"])
+            elif captcha_response.json()["status"]==1 :
                 return captcha_response.json()['request']
 
     def __del__(self):
